Implement ConsensusValidation Program
