+++ full-node/src/main/scala/co/topl/fullnode/InMemoryState.scala
@@ -2,9 +2,14 @@ package co.topl.fullnode
 
 import cats.data.NonEmptyChain
 import co.topl.models.utility.Ratio
-import co.topl.models.{BlockV2, Bytes, Evidence}
+import co.topl.models.{BlockV2, Bytes, TaktikosAddress}
 
-case class InMemoryState(tines: NonEmptyChain[Tine], relativeStake: Map[Evidence, Ratio], epochNonce: Bytes) {
+/**
+ * @param tines
+ * @param relativeStake from N-2 Epoch
+ * @param epochNonce from N-2 Epoch
+ */
+case class InMemoryState(tines: NonEmptyChain[Tine], relativeStake: Map[TaktikosAddress, Ratio], epochNonce: Bytes) {
 
   def append(nextBlock: BlockV2): InMemoryState = copy(
     NonEmptyChain.fromChainPrepend(tines.head.append(nextBlock), tines.tail)
