+++ typeclasses/src/main/scala/co/topl/typeclasses/BlockGenesis.scala
@@ -4,41 +4,43 @@ import cats.Eval
 import co.topl.models._
 import co.topl.models.utility.HasLength.implicits._
 import co.topl.models.utility.Lengths._
-import co.topl.models.utility.{Lengths, Sized}
+import co.topl.models.utility.{Length, Lengths, Sized}
 import co.topl.typeclasses.ContainsTransactions.Instances._
 import co.topl.typeclasses.ContainsTransactions.ops._
 import co.topl.typeclasses.Identifiable.Instances._
 import co.topl.typeclasses.Identifiable.ops._
 
 object BlockGenesis {
+
+  private def zeroBytes[L <: Length](implicit l: L): Sized.Strict[Bytes, L] =
+    Sized.strict[Bytes, L](Bytes(Array.fill(l.value)(0: Byte))).toOption.get
+
   val ParentId: TypedIdentifier = TypedBytes(IdentifierTypes.Block.HeaderV2, Bytes(Array.fill[Byte](32)(0)))
 
   val vrfCertificate: VrfCertificate = VrfCertificate(
-    PublicKeys.Vrf(
-      PublicKeys.Ed25519(Sized.strict[Bytes, PublicKeys.Ed25519.Length](Bytes(Array.fill(32)(0: Byte))).toOption.get)
-    ),
-    Proofs.Consensus.Nonce(Sized.strict[Bytes, Lengths.`80`.type](Bytes(Array.fill(80)(0: Byte))).toOption.get),
-    Proofs.Consensus.VrfTest(Sized.strict[Bytes, Lengths.`80`.type](Bytes(Array.fill(80)(0: Byte))).toOption.get)
+    PublicKeys.Vrf(PublicKeys.Ed25519(zeroBytes[PublicKeys.Ed25519.Length])),
+    Proofs.Consensus.Nonce(zeroBytes(Lengths.`80`)),
+    Proofs.Consensus.VrfTest(zeroBytes(Lengths.`80`))
   )
 
   val kesCertificate: KesCertificate =
     KesCertificate(
       PublicKeys.Kes(
-        Sized.strict[Bytes, PublicKeys.Kes.Length](Bytes(Array.fill(32)(0: Byte))).toOption.get,
+        zeroBytes[PublicKeys.Kes.Length],
         slot = 0
       ),
       Proofs.Consensus.KesCertificate(
-        Sized.strict[Bytes, Lengths.`64`.type](Bytes(Array.fill(64)(0: Byte))).toOption.get
+        zeroBytes(Lengths.`64`)
       ),
-      Proofs.Consensus.MMM(Sized.strict[Bytes, Lengths.`1440`.type](Bytes(Array.fill(1440)(0: Byte))).toOption.get),
+      Proofs.Consensus.MMM(zeroBytes(Lengths.`1440`)),
       slotOffset = 0
     )
 
   def apply(transactions: Seq[Transaction]): Eval[BlockV2] = Eval.later {
     val address = TaktikosAddress(
-      Sized.strict[Bytes, Lengths.`32`.type](Bytes(Array.fill[Byte](32)(0))).toOption.get,
-      Sized.strict[Bytes, Lengths.`32`.type](Bytes(Array.fill[Byte](32)(0))).toOption.get,
-      Sized.strict[Bytes, Lengths.`64`.type](Bytes(Array.fill[Byte](64)(0))).toOption.get
+      zeroBytes(Lengths.`32`),
+      zeroBytes(Lengths.`32`),
+      zeroBytes(Lengths.`64`)
     )
     val header =
       BlockHeaderV2(
